Ensure each controller only receives one router
