+++ lib/cartodb/controllers/index.js
@@ -2,7 +2,5 @@ module.exports = {
     Analyses: require('./analyses'),
     Layergroup: require('./layergroup'),
     Map: require('./map'),
-    NamedMaps: require('./named_maps'),
-    NamedMapsAdmin: require('./named_maps_admin'),
     ServerInfo: require('./server_info')
 };
