+++ lib/cartodb/controllers/template/admin.js
@@ -1,9 +1,9 @@
-const { templateName } = require('../backends/template_maps');
-const cors = require('../middleware/cors');
-const credentials = require('../middleware/credentials');
-const rateLimit = require('../middleware/rate-limit');
+const { templateName } = require('../../backends/template_maps');
+const cors = require('../../middleware/cors');
+const credentials = require('../../middleware/credentials');
+const rateLimit = require('../../middleware/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
-const sendResponse = require('../middleware/send-response');
+const sendResponse = require('../../middleware/send-response');
 
 /**
  * @param {AuthApi} authApi
@@ -11,21 +11,21 @@ const sendResponse = require('../middleware/send-response');
  * @param {TemplateMaps} templateMaps
  * @constructor
  */
-function NamedMapsAdminController(authApi, templateMaps, userLimitsApi) {
+function AdminTemplateController(authApi, templateMaps, userLimitsApi) {
     this.authApi = authApi;
     this.templateMaps = templateMaps;
     this.userLimitsApi = userLimitsApi;
 }
 
-module.exports = NamedMapsAdminController;
+module.exports = AdminTemplateController;
 
-NamedMapsAdminController.prototype.register = function (templateRouter) {
+AdminTemplateController.prototype.register = function (templateRouter) {
     templateRouter.post(
         `/`,
         credentials(),
-        checkContentType({ action: 'POST', label: 'POST TEMPLATE' }),
         authorizedByAPIKey({ authApi: this.authApi, action: 'create', label: 'POST TEMPLATE' }),
         rateLimit(this.userLimitsApi, RATE_LIMIT_ENDPOINTS_GROUPS.NAMED_CREATE),
+        checkContentType({ action: 'POST', label: 'POST TEMPLATE' }),
         createTemplate({ templateMaps: this.templateMaps }),
         sendResponse()
     );
@@ -33,9 +33,9 @@ NamedMapsAdminController.prototype.register = function (templateRouter) {
     templateRouter.put(
         `/:template_id`,
         credentials(),
-        checkContentType({ action: 'PUT', label: 'PUT TEMPLATE' }),
         authorizedByAPIKey({ authApi: this.authApi, action: 'update', label: 'PUT TEMPLATE' }),
         rateLimit(this.userLimitsApi, RATE_LIMIT_ENDPOINTS_GROUPS.NAMED_UPDATE),
+        checkContentType({ action: 'PUT', label: 'PUT TEMPLATE' }),
         updateTemplate({ templateMaps: this.templateMaps }),
         sendResponse()
     );
