+++ lib/cartodb/controllers/template/preview.js
@@ -1,14 +1,14 @@
-const cleanUpQueryParams = require('../middleware/clean-up-query-params');
-const credentials = require('../middleware/credentials');
-const dbConnSetup = require('../middleware/db-conn-setup');
-const authorize = require('../middleware/authorize');
-const cacheControlHeader = require('../middleware/cache-control-header');
-const cacheChannelHeader = require('../middleware/cache-channel-header');
-const surrogateKeyHeader = require('../middleware/surrogate-key-header');
-const lastModifiedHeader = require('../middleware/last-modified-header');
-const sendResponse = require('../middleware/send-response');
-const vectorError = require('../middleware/vector-error');
-const rateLimit = require('../middleware/rate-limit');
+const cleanUpQueryParams = require('../../middleware/clean-up-query-params');
+const credentials = require('../../middleware/credentials');
+const dbConnSetup = require('../../middleware/db-conn-setup');
+const authorize = require('../../middleware/authorize');
+const namedMapProvider = require('./middlewares/named-map-provider');
+const cacheControlHeader = require('../../middleware/cache-control-header');
+const cacheChannelHeader = require('../../middleware/cache-channel-header');
+const surrogateKeyHeader = require('../../middleware/surrogate-key-header');
+const lastModifiedHeader = require('../../middleware/last-modified-header');
+const sendResponse = require('../../middleware/send-response');
+const rateLimit = require('../../middleware/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
 
 const DEFAULT_ZOOM_CENTER = {
@@ -23,9 +23,8 @@ function numMapper(n) {
     return +n;
 }
 
-function NamedMapsController (
+function PreviewTemplateController (
     namedMapProviderCache,
-    tileBackend,
     previewBackend,
     surrogateKeysCache,
     tablesExtentApi,
@@ -35,7 +34,6 @@ function NamedMapsController (
     userLimitsApi
 ) {
     this.namedMapProviderCache = namedMapProviderCache;
-    this.tileBackend = tileBackend;
     this.previewBackend = previewBackend;
     this.surrogateKeysCache = surrogateKeysCache;
     this.tablesExtentApi = tablesExtentApi;
@@ -45,34 +43,9 @@ function NamedMapsController (
     this.userLimitsApi = userLimitsApi;
 }
 
-module.exports = NamedMapsController;
+module.exports = PreviewTemplateController;
 
-NamedMapsController.prototype.register = function(mapRouter, templateRouter) {
-
-    templateRouter.get(
-        `/:template_id/:layer/:z/:x/:y.(:format)`,
-        credentials(),
-        authorize(this.authApi),
-        dbConnSetup(this.pgConnection),
-        rateLimit(this.userLimitsApi, RATE_LIMIT_ENDPOINTS_GROUPS.NAMED_TILES),
-        cleanUpQueryParams(),
-        getNamedMapProvider({
-            namedMapProviderCache: this.namedMapProviderCache,
-            label: 'NAMED_MAP_TILE'
-        }),
-        getTile({
-            tileBackend: this.tileBackend,
-            label: 'NAMED_MAP_TILE'
-        }),
-        setContentTypeHeader(),
-        cacheControlHeader(),
-        cacheChannelHeader(),
-        surrogateKeyHeader({ surrogateKeysCache: this.surrogateKeysCache }),
-        lastModifiedHeader(),
-        sendResponse(),
-        vectorError()
-    );
-
+PreviewTemplateController.prototype.register = function (mapRouter) {
     mapRouter.get(
         `/static/named/:template_id/:width/:height.:format`,
         credentials(),
@@ -80,7 +53,7 @@ NamedMapsController.prototype.register = function(mapRouter, templateRouter) {
         dbConnSetup(this.pgConnection),
         rateLimit(this.userLimitsApi, RATE_LIMIT_ENDPOINTS_GROUPS.STATIC_NAMED),
         cleanUpQueryParams(['layer', 'zoom', 'lon', 'lat', 'bbox']),
-        getNamedMapProvider({
+        namedMapProvider({
             namedMapProviderCache: this.namedMapProviderCache,
             label: 'STATIC_VIZ_MAP', forcedFormat: 'png'
         }),
@@ -100,39 +73,6 @@ NamedMapsController.prototype.register = function(mapRouter, templateRouter) {
         sendResponse()
     );
 };
-
-function getNamedMapProvider ({ namedMapProviderCache, label, forcedFormat = null }) {
-    return function getNamedMapProviderMiddleware (req, res, next) {
-        const { user, token, cache_buster, api_key } = res.locals;
-        const { dbuser, dbname, dbpassword, dbhost, dbport } = res.locals;
-        const { template_id, layer: layerFromParams, z, x, y, format } = req.params;
-        const { layer: layerFromQuery } = req.query;
-
-        const params = {
-            user, token, cache_buster, api_key,
-            dbuser, dbname, dbpassword, dbhost, dbport,
-            template_id, layer: (layerFromQuery || layerFromParams), z, x, y, format
-        };
-
-        if (forcedFormat) {
-            params.format = forcedFormat;
-            params.layer = params.layer || 'all';
-        }
-
-        const { config, auth_token } = req.query;
-
-        namedMapProviderCache.get(user, template_id, config, auth_token, params, (err, namedMapProvider) => {
-            if (err) {
-                err.label = label;
-                return next(err);
-            }
-
-            res.locals.mapConfigProvider = namedMapProvider;
-
-            next();
-        });
-    };
-}
 
 function getTemplate ({ label }) {
     return function getTemplateMiddleware (req, res, next) {
@@ -194,32 +134,6 @@ function prepareLayerFilterFromPreviewLayers ({ namedMapProviderCache, label }) 
             }
 
             res.locals.mapConfigProvider = provider;
-
-            next();
-        });
-    };
-}
-
-function getTile ({ tileBackend, label }) {
-    return function getTileMiddleware (req, res, next) {
-        const { mapConfigProvider } = res.locals;
-        const { layer, z, x, y, format } = req.params;
-        const params = { layer, z, x, y, format };
-
-        tileBackend.getTile(mapConfigProvider, params, (err, tile, headers, stats) => {
-            req.profiler.add(stats);
-            req.profiler.done('render-' + format);
-
-            if (err) {
-                err.label = label;
-                return next(err);
-            }
-
-            if (headers) {
-                res.set(headers);
-            }
-
-            res.body = tile;
 
             next();
         });
