+++ lib/cartodb/server.js
@@ -7,6 +7,9 @@ var cartodbRedis = require('cartodb-redis');
 var _ = require('underscore');
 
 var controller = require('./controllers');
+const AdminTemplateController = require('./controllers/template/admin');
+const PreviewTemplateController = require('./controllers/template/preview');
+const TileTemplateController = require('./controllers/template/tile');
 
 var SurrogateKeysCache = require('./cache/surrogate_keys_cache');
 var NamedMapsCacheEntry = require('./cache/model/named_maps_entry');
@@ -224,8 +227,6 @@ module.exports = function(serverOptions) {
     const templateRouter = express.Router();
     const monitorRouter = express.Router();
 
-    const { base_url_mapconfig: mapConfigBasePath, base_url_templated: templateBasePath } = serverOptions;
-
     new controller.Layergroup(
         pgConnection,
         mapStore,
@@ -252,9 +253,17 @@ module.exports = function(serverOptions) {
         authApi
     ).register(mapRouter, templateRouter);
 
-    new controller.NamedMaps(
+    new TileTemplateController(
         namedMapProviderCache,
         tileBackend,
+        surrogateKeysCache,
+        pgConnection,
+        authApi,
+        userLimitsApi
+    ).register(templateRouter);
+
+    new PreviewTemplateController(
+        namedMapProviderCache,
         previewBackend,
         surrogateKeysCache,
         tablesExtentApi,
@@ -262,13 +271,19 @@ module.exports = function(serverOptions) {
         pgConnection,
         authApi,
         userLimitsApi
-    ).register(mapRouter, templateRouter);
+    ).register(mapRouter);
 
-    new controller.NamedMapsAdmin(authApi, templateMaps, userLimitsApi).register(templateRouter);
+    new AdminTemplateController(
+        authApi,
+        templateMaps,
+        userLimitsApi
+    ).register(templateRouter);
 
     new controller.Analyses(pgConnection, authApi, userLimitsApi).register(mapRouter);
 
     new controller.ServerInfo(versions).register(monitorRouter);
+
+    const { base_url_mapconfig: mapConfigBasePath, base_url_templated: templateBasePath } = serverOptions;
 
     app.use(mapConfigBasePath, mapRouter);
     app.use(templateBasePath, templateRouter);
